@@ -32,21 +38,34 @@
         }
     ],
     "require": {
-        "php": ">=5.5.0",
-        "ext-mbstring": "*",
-        "ext-mysqli": "*",
-        "ext-xml": "*",
-        "ext-pcre": "*",
+        "php": "^7.1.3 || ^8.0",
+        "ext-hash": "*",
+        "ext-iconv": "*",
         "ext-json": "*",
-        "phpmyadmin/sql-parser": "^4.2.3",
-        "phpmyadmin/motranslator": "^3.0",
+        "ext-mysqli": "*",
+        "ext-pcre": "*",
+        "ext-xml": "*",
+        "google/recaptcha": "^1.1",
+        "nikic/fast-route": "^1.3",
+        "phpmyadmin/motranslator": "^5.0",
         "phpmyadmin/shapefile": "^2.0",
-        "tecnickcom/tcpdf": "^6.2",
+        "phpmyadmin/sql-parser": "^5.0",
+        "phpmyadmin/twig-i18n-extension": "^3.0",
         "phpseclib/phpseclib": "^2.0",
-        "google/recaptcha": "^1.1"
+        "symfony/config": "^4.4.9",
+        "symfony/dependency-injection": "^4.4.9",
+        "symfony/expression-language": "^4.4.9",
+        "symfony/polyfill-ctype": "^1.17.0",
+        "symfony/polyfill-mbstring": "^1.17.0",
+        "twig/twig": "^2.9 || ^3",
+        "williamdes/mariadb-mysql-kbs": "^1.2"
     },
     "conflict": {
-        "phpseclib/phpseclib": "2.0.8"
+        "phpseclib/phpseclib": "2.0.8",
+        "tecnickcom/tcpdf": "<6.2",
+        "pragmarx/google2fa": "<6.1.0 || >8.0",
+        "pragmarx/google2fa-qrcode": "<1.0.1",
+        "samyoul/u2f-php-server": "<1.1"
     },
     "suggest": {
         "ext-openssl": "Cookie encryption",
