@@ -1,30 +1,37 @@
 <?php
-/* vim: set expandtab sw=4 ts=4 sts=4: */
 /**
  * Validation callback.
- *
- * @package PhpMyAdmin-Setup
  */
+
+declare(strict_types=1);
+
+use PhpMyAdmin\Config\Validator;
+use PhpMyAdmin\Core;
+
+if (! defined('ROOT_PATH')) {
+    // phpcs:disable PSR1.Files.SideEffects
+    define('ROOT_PATH', dirname(__DIR__) . DIRECTORY_SEPARATOR);
+    // phpcs:enable
+}
 
 /**
  * Core libraries.
  */
-require './lib/common.inc.php';
+require ROOT_PATH . 'setup/lib/common.inc.php';
 
-$validators = array();
-require './libraries/config/Validator.php';
+$validators = [];
 
-PMA_headerJSON();
+Core::headerJSON();
 
-$ids = PMA_isValid($_POST['id'], 'scalar') ? $_POST['id'] : null;
+$ids = Core::isValid($_POST['id'], 'scalar') ? $_POST['id'] : null;
 $vids = explode(',', $ids);
-$vals = PMA_isValid($_POST['values'], 'scalar') ? $_POST['values'] : null;
+$vals = Core::isValid($_POST['values'], 'scalar') ? $_POST['values'] : null;
 $values = json_decode($vals);
-if (!($values instanceof stdClass)) {
-    PMA_fatalError(__('Wrong data'));
+if (! ($values instanceof stdClass)) {
+    Core::fatalError(__('Wrong data'));
 }
-$values = (array)$values;
-$result = PMA\libraries\config\Validator::validate($GLOBALS['ConfigFile'], $vids, $values, true);
+$values = (array) $values;
+$result = Validator::validate($GLOBALS['ConfigFile'], $vids, $values, true);
 if ($result === false) {
     $result = sprintf(
         __('Wrong data or no validation for %s'),
