@@ -1,41 +1,54 @@
 <?php
-/* vim: set expandtab sw=4 ts=4 sts=4: */
 /**
- * File for vendor customisation, you can change here paths or some behaviour,
+ * File for vendor customization, you can change here paths or some behaviour,
  * which vendors such as Linux distributions might want to change.
  *
  * For changing this file you should know what you are doing. For this reason
  * options here are not part of normal configuration.
- *
- * @package PhpMyAdmin
  */
+
+declare(strict_types=1);
+
+// phpcs:disable PSR1.Files.SideEffects
 if (! defined('PHPMYADMIN')) {
     exit;
 }
+// phpcs:enable
+
+/**
+ * Path to vendor autoload file. Useful when you want to
+ * have have vendor dependencies somewhere else.
+ */
+define('AUTOLOAD_FILE', ROOT_PATH . 'vendor' . DIRECTORY_SEPARATOR . 'autoload.php');
+
+/**
+ * Directory where cache files are stored.
+ */
+define('TEMP_DIR', ROOT_PATH . 'tmp' . DIRECTORY_SEPARATOR);
 
 /**
  * Path to changelog file, can be gzip compressed. Useful when you want to
  * have documentation somewhere else, eg. /usr/share/doc.
  */
-define('CHANGELOG_FILE', './ChangeLog');
+define('CHANGELOG_FILE', ROOT_PATH . 'ChangeLog');
 
 /**
  * Path to license file. Useful when you want to have documentation somewhere
  * else, eg. /usr/share/doc.
  */
-define('LICENSE_FILE', './LICENSE');
+define('LICENSE_FILE', ROOT_PATH . 'LICENSE');
 
 /**
  * Directory where SQL scripts to create/upgrade configuration storage reside.
  */
-define('SQL_DIR', './sql/');
+define('SQL_DIR', ROOT_PATH . 'sql' . DIRECTORY_SEPARATOR);
 
 /**
  * Directory where configuration files are stored.
  * It is not used directly in code, just a convenient
  * define used further in this file.
  */
-define('CONFIG_DIR', '');
+define('CONFIG_DIR', ROOT_PATH);
 
 /**
  * Filename of a configuration file.
@@ -60,10 +73,9 @@ define('VERSION_CHECK_DEFAULT', true);
 /**
  * Path to files with compiled locales (*.mo)
  */
-define('LOCALE_PATH', './locale/');
+define('LOCALE_PATH', ROOT_PATH . 'locale' . DIRECTORY_SEPARATOR);
 
 /**
- * Avoid referring to nonexistent files (causes warnings when open_basedir
- * is used)
+ * Define the cache directory for routing cache an other cache files
  */
-define('K_PATH_IMAGES', '');
+define('CACHE_DIR', ROOT_PATH . 'libraries' . DIRECTORY_SEPARATOR . 'cache' . DIRECTORY_SEPARATOR);
