@@ -1,50 +1,53 @@
 <?php
-/* vim: set expandtab sw=4 ts=4 sts=4: */
 /**
  * Front controller for config view / download and clear
- *
- * @package PhpMyAdmin-Setup
  */
-use PMA\libraries\config\FormDisplay;
-use PMA\setup\lib\ConfigGenerator;
-use PMA\libraries\URL;
+
+declare(strict_types=1);
+
+use PhpMyAdmin\Config\Forms\Setup\ConfigForm;
+use PhpMyAdmin\Core;
+use PhpMyAdmin\Response;
+use PhpMyAdmin\Setup\ConfigGenerator;
+use PhpMyAdmin\Url;
+
+if (! defined('ROOT_PATH')) {
+    // phpcs:disable PSR1.Files.SideEffects
+    define('ROOT_PATH', dirname(__DIR__) . DIRECTORY_SEPARATOR);
+    // phpcs:enable
+}
 
 /**
  * Core libraries.
  */
-require './lib/common.inc.php';
+require ROOT_PATH . 'setup/lib/common.inc.php';
 
-require './libraries/config/setup.forms.php';
+$form_display = new ConfigForm($GLOBALS['ConfigFile']);
+$form_display->save('Config');
 
-$form_display = new FormDisplay($GLOBALS['ConfigFile']);
-$form_display->registerForm('_config.php', $forms['_config.php']);
-$form_display->save('_config.php');
+$response = Response::getInstance();
+$response->disable();
 
 if (isset($_POST['eol'])) {
-    $_SESSION['eol'] = ($_POST['eol'] == 'unix') ? 'unix' : 'win';
+    $_SESSION['eol'] = $_POST['eol'] === 'unix' ? 'unix' : 'win';
 }
 
-if (PMA_ifSetOr($_POST['submit_clear'], '')) {
-    //
+if (Core::ifSetOr($_POST['submit_clear'], '')) {
     // Clear current config and return to main page
-    //
     $GLOBALS['ConfigFile']->resetConfigData();
     // drop post data
-    header('HTTP/1.1 303 See Other');
-    header('Location: index.php' . URL::getCommonRaw());
-    exit;
-} elseif (PMA_ifSetOr($_POST['submit_download'], '')) {
-    //
-    // Output generated config file
-    //
-    PMA_downloadHeader('config.inc.php', 'text/plain');
-    echo ConfigGenerator::getConfigFile($GLOBALS['ConfigFile']);
-    exit;
-} else {
-    //
-    // Show generated config file in a <textarea>
-    //
-    header('HTTP/1.1 303 See Other');
-    header('Location: index.php' . URL::getCommonRaw(array('page' => 'config')));
+    $response->generateHeader303('index.php' . Url::getCommonRaw());
     exit;
 }
+
+if (Core::ifSetOr($_POST['submit_download'], '')) {
+    // Output generated config file
+    Core::downloadHeader('config.inc.php', 'text/plain');
+    $response->disable();
+    echo ConfigGenerator::getConfigFile($GLOBALS['ConfigFile']);
+    exit;
+}
+
+// Show generated config file in a <textarea>
+$response->generateHeader303('index.php' . Url::getCommonRaw(['page' => 'config']));
+exit;
