@@ -1,65 +1,82 @@
 <?php
-/* vim: set expandtab sw=4 ts=4 sts=4: */
 /**
  * Front controller for setup script
- *
- * @package PhpMyAdmin-Setup
- * @license https://www.gnu.org/licenses/gpl.html GNU GPL 2.0
  */
 
-/**
- * Core libraries.
- */
-require './lib/common.inc.php';
+declare(strict_types=1);
 
-if (@file_exists(CONFIG_FILE) && ! $cfg['DBG']['demo']) {
-    PMA_fatalError(__('Configuration already exists, setup is disabled!'));
+use PhpMyAdmin\Controllers\Setup\ConfigController;
+use PhpMyAdmin\Controllers\Setup\FormController;
+use PhpMyAdmin\Controllers\Setup\HomeController;
+use PhpMyAdmin\Controllers\Setup\ServersController;
+use PhpMyAdmin\Core;
+use PhpMyAdmin\Template;
+use PhpMyAdmin\Url;
+
+if (! defined('ROOT_PATH')) {
+    // phpcs:disable PSR1.Files.SideEffects
+    define('ROOT_PATH', dirname(__DIR__) . DIRECTORY_SEPARATOR);
+    // phpcs:enable
 }
 
-$page = PMA_isValid($_GET['page'], 'scalar') ? $_GET['page'] : null;
+global $cfg;
+
+require ROOT_PATH . 'setup/lib/common.inc.php';
+
+if (@file_exists(CONFIG_FILE) && ! $cfg['DBG']['demo']) {
+    Core::fatalError(__('Configuration already exists, setup is disabled!'));
+}
+
+$page = Core::isValid($_GET['page'], 'scalar') ? (string) $_GET['page'] : '';
 $page = preg_replace('/[^a-z]/', '', $page);
 if ($page === '') {
     $page = 'index';
 }
-if (!@file_exists("./setup/frames/$page.inc.php")) {
-    // it will happen only when entering URL by hand, we don't care for these cases
-    PMA_fatalError(__('Wrong GET file attribute value'));
+
+Core::noCacheHeader();
+
+if ($page === 'form') {
+    $controller = new FormController($GLOBALS['ConfigFile'], new Template());
+    echo $controller->index([
+        'formset' => $_GET['formset'] ?? null,
+    ]);
+
+    return;
 }
 
-// Handle done action info
-$action_done = PMA_isValid($_GET['action_done'], 'scalar') ? $_GET['action_done'] : null;
-$action_done = preg_replace('/[^a-z_]/', '', $action_done);
+if ($page === 'config') {
+    $controller = new ConfigController($GLOBALS['ConfigFile'], new Template());
+    echo $controller->index([
+        'formset' => $_GET['formset'] ?? null,
+        'eol' => $_GET['eol'] ?? null,
+    ]);
 
-PMA_noCacheHeader();
+    return;
+}
 
-?>
-<!DOCTYPE HTML>
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head>
-<meta charset="utf-8" />
-<title>phpMyAdmin setup</title>
-<link href="../favicon.ico" rel="icon" type="image/x-icon" />
-<link href="../favicon.ico" rel="shortcut icon" type="image/x-icon" />
-<link href="styles.css" rel="stylesheet" type="text/css" />
-<script type="text/javascript" src="../js/jquery/jquery.min.js"></script>
-<script type="text/javascript" src="../js/jquery/jquery-ui.min.js">
-</script>
-<script type="text/javascript" src="ajax.js"></script>
-<script type="text/javascript" src="../js/config.js"></script>
-<script type="text/javascript" src="scripts.js"></script>
-<script type="text/javascript" src="../js/messages.php"></script>
-</head>
-<body>
-<h1><span class="blue">php</span><span class="orange">MyAdmin</span>  setup</h1>
-<div id="menu">
-<?php
-require './setup/frames/menu.inc.php';
-?>
-</div>
-<div id="page">
-<?php
-require "./setup/frames/$page.inc.php";
-?>
-</div>
-</body>
-</html>
+if ($page === 'servers') {
+    $controller = new ServersController($GLOBALS['ConfigFile'], new Template());
+    if (isset($_GET['mode']) && $_GET['mode'] === 'remove' && $_SERVER['REQUEST_METHOD'] === 'POST') {
+        $controller->destroy([
+            'id' => $_GET['id'] ?? null,
+        ]);
+        header('Location: index.php' . Url::getCommonRaw());
+
+        return;
+    }
+
+    echo $controller->index([
+        'formset' => $_GET['formset'] ?? null,
+        'mode' => $_GET['mode'] ?? null,
+        'id' => $_GET['id'] ?? null,
+    ]);
+
+    return;
+}
+
+$controller = new HomeController($GLOBALS['ConfigFile'], new Template());
+echo $controller->index([
+    'formset' => $_GET['formset'] ?? null,
+    'action_done' => $_GET['action_done'] ?? null,
+    'version_check' => $_GET['version_check'] ?? null,
+]);
