@@ -1,38 +1,35 @@
 <?php
-/* vim: set expandtab sw=4 ts=4 sts=4: */
 /**
  * Loads libraries/common.inc.php and preforms some additional actions
- *
- * @package PhpMyAdmin-Setup
  */
-use PMA\libraries\config\ConfigFile;
 
-/**
- * Do not include full common.
- * @ignore
- */
+declare(strict_types=1);
+
+use PhpMyAdmin\Config\ConfigFile;
+use PhpMyAdmin\DatabaseInterface;
+
+// phpcs:disable PSR1.Files.SideEffects
 define('PMA_MINIMUM_COMMON', true);
-define('PMA_SETUP', true);
+// phpcs:enable
+
 chdir('..');
 
-if (!file_exists('./libraries/common.inc.php')) {
+if (! file_exists(ROOT_PATH . 'libraries/common.inc.php')) {
     die('Bad invocation!');
 }
 
-require_once './libraries/common.inc.php';
-require_once './libraries/config/config_functions.lib.php';
-require_once './libraries/config/messages.inc.php';
-require_once './libraries/user_preferences.lib.php';
+require_once ROOT_PATH . 'libraries/common.inc.php';
 
 // use default error handler
 restore_error_handler();
 
 // Save current language in a cookie, required since we use PMA_MINIMUM_COMMON
-$GLOBALS['PMA_Config']->setCookie('pma_lang', $GLOBALS['lang']);
+$GLOBALS['PMA_Config']->setCookie('pma_lang', (string) $GLOBALS['lang']);
+$GLOBALS['PMA_Config']->set('is_setup', true);
 
 $GLOBALS['ConfigFile'] = new ConfigFile();
 $GLOBALS['ConfigFile']->setPersistKeys(
-    array(
+    [
         'DefaultLang',
         'ServerDefault',
         'UploadDir',
@@ -43,10 +40,11 @@ $GLOBALS['ConfigFile']->setPersistKeys(
         'Servers/1/socket',
         'Servers/1/auth_type',
         'Servers/1/user',
-        'Servers/1/password'
-    )
+        'Servers/1/password',
+    ]
 );
+
+$GLOBALS['dbi'] = DatabaseInterface::load();
 
 // allows for redirection even after sending some data
 ob_start();
-
