@@ -1,10 +1,15 @@
 <?php
-/* vim: set expandtab sw=4 ts=4 sts=4: */
 /**
  * Simple wrapper just to enable error reporting and include config
- *
- * @package PhpMyAdmin
  */
+
+declare(strict_types=1);
+
+if (! defined('ROOT_PATH')) {
+    // phpcs:disable PSR1.Files.SideEffects
+    define('ROOT_PATH', __DIR__ . DIRECTORY_SEPARATOR);
+    // phpcs:enable
+}
 
 // rfc2616 - Section 14.21
 header('Expires: ' . gmdate(DATE_RFC1123));
@@ -14,9 +19,8 @@ header(
     . '  pre-check=0, post-check=0, max-age=0'
 );
 if (isset($_SERVER['HTTP_USER_AGENT'])
-    && stristr($_SERVER['HTTP_USER_AGENT'], 'MSIE')
+    && stripos($_SERVER['HTTP_USER_AGENT'], 'MSIE') !== false
 ) {
-
     /* FIXME: Why is this special case for IE needed? */
     header('Pragma: public');
 } else {
@@ -28,9 +32,17 @@ if (isset($_SERVER['HTTP_USER_AGENT'])
 }
 header('Content-Type: text/html; charset=utf-8');
 
-require 'libraries/vendor_config.php';
+// phpcs:disable PSR1.Files.SideEffects
+define('PHPMYADMIN', true);
+// phpcs:enable
 
-error_reporting(E_ALL);
+require ROOT_PATH . 'libraries/vendor_config.php';
+
+// issue #16256 - This only works with php 8.0+
+if (function_exists('error_reporting')) {
+    error_reporting(E_ALL);
+}
+
 /**
  * Read config file.
  */
