@@ -1,41 +1,54 @@
 <?php
-/* vim: set expandtab sw=4 ts=4 sts=4: */
 /**
  * URL redirector to avoid leaking Referer with some sensitive information.
- *
- * @package PhpMyAdmin
  */
-use PMA\libraries\Sanitize;
-use PMA\libraries\Response;
 
-/**
- * Gets core libraries and defines some variables
- */
+declare(strict_types=1);
+
+use PhpMyAdmin\Core;
+use PhpMyAdmin\DatabaseInterface;
+use PhpMyAdmin\Response;
+use PhpMyAdmin\Sanitize;
+
+if (! defined('ROOT_PATH')) {
+    // phpcs:disable PSR1.Files.SideEffects
+    define('ROOT_PATH', __DIR__ . DIRECTORY_SEPARATOR);
+    // phpcs:enable
+}
+
+global $containerBuilder, $dbi;
+
+// phpcs:disable PSR1.Files.SideEffects
 define('PMA_MINIMUM_COMMON', true);
-require_once './libraries/common.inc.php';
+// phpcs:enable
+
+require_once ROOT_PATH . 'libraries/common.inc.php';
+
+// Load database service because services.php is not available here
+$dbi = DatabaseInterface::load();
+$containerBuilder->set(DatabaseInterface::class, $dbi);
 
 // Only output the http headers
 $response = Response::getInstance();
 $response->getHeader()->sendHttpHeaders();
 $response->disable();
 
-if (! PMA_isValid($_REQUEST['url'])
-    || ! preg_match('/^https:\/\/[^\n\r]*$/', $_REQUEST['url'])
-    || ! PMA_isAllowedDomain($_REQUEST['url'])
+if (! Core::isValid($_GET['url'])
+    || ! preg_match('/^https:\/\/[^\n\r]*$/', $_GET['url'])
+    || ! Core::isAllowedDomain($_GET['url'])
 ) {
-    PMA_sendHeaderLocation('./');
+    Core::sendHeaderLocation('./');
 } else {
     // JavaScript redirection is necessary. Because if header() is used
     //  then web browser sometimes does not change the HTTP_REFERER
     //  field and so with old URL as Referer, token also goes to
     //  external site.
-    echo "<script type='text/javascript'>
-            window.onload=function(){
-                window.location='" , Sanitize::escapeJsString($_REQUEST['url']) , "';
-            }
-        </script>";
+    $template = $containerBuilder->get('template');
+    echo $template->render('javascript/redirect', [
+        'url' => Sanitize::escapeJsString($_GET['url']),
+    ]);
     // Display redirecting msg on screen.
-    // Do not display the value of $_REQUEST['url'] to avoid showing injected content
+    // Do not display the value of $_GET['url'] to avoid showing injected content
     echo __('Taking you to the target site.');
 }
-die();
+die;
